My Cars For Users Is Complete/Rent btn Is Complete

+++ MyCarManagmentProject/Person.cs
@@ -15,7 +15,7 @@ namespace MyCarManagmentProject
        
         public string LastName { get;  set; }
  
-        public int WalletBalance { get;  set; }
+        public decimal WalletBalance { get;  set; }
         public string UserName { get; set; }
         public string Password { get;  set; }
         public int Id { get; set; }
